+++ src/hooks/usePersonalData/localStorageClient/localStorageClient.tsx
@@ -1,3 +1,13 @@
+import {
+  CompletedVideo,
+  COMPLETED_VIDEO,
+  FollowedChannel,
+  InterruptedVideo,
+  INTERRUPTED_VIDEO,
+  PersonalDataClient,
+  WatchedVideo,
+} from './types'
+
 const promisify = <T,>(fn: (...args: unknown[]) => T) => (...args: Parameters<typeof fn>) =>
   new Promise((resolve) => resolve(fn(...args))) as Promise<T>
 
@@ -111,58 +121,3 @@ const localStorageClient: PersonalDataClient = {
 }
 
 export default localStorageClient
-
-type WatchedVideoFields = {
-  id: string
-}
-
-export const INTERRUPTED_VIDEO = 'INTERRUPTED'
-export const COMPLETED_VIDEO = 'COMPLETED'
-
-type InterruptedVideo = WatchedVideoFields & {
-  __typename: typeof INTERRUPTED_VIDEO
-  timestamp: number
-}
-type CompletedVideo = WatchedVideoFields & {
-  __typename: typeof COMPLETED_VIDEO
-  id: string
-}
-export type WatchedVideo = InterruptedVideo | CompletedVideo
-
-export type FollowedChannel = {
-  id: string
-}
-
-export interface PersonalDataClient {
-  // ==== watched videos ====
-
-  // get all the interrupted or completed videos
-  watchedVideos: () => Promise<WatchedVideo[]>
-
-  // get all the interrupted videos
-  interruptedVideos: () => Promise<InterruptedVideo[]>
-
-  // get all the completed videos
-  completedVideos: () => Promise<CompletedVideo[]>
-
-  // get status of one specific watched video
-  watchedVideo: (id: string) => Promise<WatchedVideo | null>
-
-  // mark the video as interrupted or completed
-  setWatchedVideo: (
-    __typename: typeof INTERRUPTED_VIDEO | typeof COMPLETED_VIDEO,
-    id: string,
-    timestamp?: number
-  ) => Promise<void>
-
-  // ==== followed channels ====
-
-  // get all the followed channels
-  followedChannels: () => Promise<FollowedChannel[]>
-
-  // set the channel following status
-  setChannelFollowing: (id: string, follow: boolean) => Promise<void>
-
-  // get the following status of one specific channel
-  isFollowingChannel: (id: string) => Promise<boolean>
-}
